Prepend http path to relative redirect location
It is possible to redirect an HTTP request to a relative path. In order to
handle this, the current http path without the part after the last slash should
be prepended to the redirect location. A path is relative when the path does
not start with "/", and does not start with "http://" or "https://".

Without this fix, siege will request the relative path from the server. This
will result in 404 errors because siege requests the wrong path.